BaseVeGenerator: now with factory, better separation between VE generator and VE generation alg
